fix: `ftm::Debug` impl fails to compile
The `ftm::Debug` impl fails to compile unless the `alloc` feature is on.
The underlying math lib allows `alloc`-free hex output, except it's not
fixed-width like the current impl does.

This technically changes the debug output format, but it should be worth
it for getting to avoid requiring `alloc` for just debug printing.

Also adds a GitHub Actions workflow to check that all feature
combinations would compile correctly so that this doesn't happen again.